refactor: move statistical functions to separate class
Try to split mathematical core operations from functions with statistical character.
